Use Stackdriver compatible JSON logs
This implementation is adapted from:
https://github.com/spring-cloud/spring-cloud-gcp/blob/b5c1fc456fb5af710b5d91c8a40c0d1642eaa9a7/spring-cloud-gcp-autoconfigure/src/main/java/org/springframework/cloud/gcp/autoconfigure/logging/StackdriverJsonLayout.java

Ideally we would use the Spring GCP logging support:
https://docs.spring.io/spring-cloud-gcp/docs/1.1.0.M1/reference/html/_stackdriver_logging_support.html#_stackdriver_logging_support

But this ends up pulling in a bunch of GCP autoconfiguration which we
don't want.

There's an open issue (& PR) to split the logging adapter into
it's own artifact:
https://github.com/spring-cloud/spring-cloud-gcp/issues/1448
https://github.com/spring-cloud/spring-cloud-gcp/pull/1455

Once this is merged & released we can switch to the official GCP
implementation.