Implement renewAccess link for expired user
This introduces a shim to enable the teacher app to determine whether
the user has 'expired' or not.

This has some downsides in that it would still be possible for the user
to access video endpoints, etc even when they do not have access to the
platform. In order to secure the platform we'd need to add additional
checks to the various microservices that we use in order to prevent
access based on the User::hasAccess function.

Ideally the expiry would be modelled in here, and the identity provider
(Keycloak) would provide us with the ability to determine - at login
time - whether the user should have access or not. Unfortunately at this
point we don't have enough knowledge of the Open ID Connect process to
implement this, but hopefully we'll do so in the future.

At this stage we're protecting against teachers gaining free access to
our platform.

[#169737339]