Account resource returns a self link
The self link is expected by e2e tests setup. It didn't fail on
Concourse because the tests are executed from a clean slate each time
over there, but when you try running the test locally against a test
database that already has some data in it then a branch that expects the
self link would be executed.