Import email from keycloak
And use it as the user's email if the username is not a valid email. We
are doing this becuase the MOE UAE SSO users don't have their emails
saved in the keycloak identity username field. We could not find a way
to override that field so we are using their email field from keycloak.

[#172814030]

Signed-off-by: Alex Glen <alex@boclips.com>